Begin work to support set-mark-to-object
