+++ Cargo.toml
@@ -10,7 +10,7 @@ rustc-serialize = "0.2.12"
 
 [dependencies.gapbuffer]
 git = "https://github.com/dlaronson/gapbuffer"
-version = "0.0.4"
+version = "0.0.5"
 
 [lib]
 name = "iota"
